Add No Data Found Check & Some Other Fixes

+++ lib/view/home_screen/widgets/restaurants_details_top_widget.dart
@@ -154,7 +154,8 @@ class RestaurantDetailTopScreen extends StatelessWidget {
                                               null
                                           ? 0
                                           : controller.detailsRestaurantList
-                                              .value.rating!,
+                                              .value.rating!
+                                              .toDouble(),
                                       // minRating: 1,
                                       direction: Axis.horizontal,
                                       itemCount: 5,
